fix(simd): correct SVE2 moment FCVT lane mapping + float-square moment SIMD tails
Two float_moment SIMD bugs from the 2026-06-27 bug-hunt sweep (subsystem: simd).

1. SVE2 moment widening (wrong math). compute_1st/2nd_moment_sve2
   (core/src/feature/arm64/moment_sve2.c) stepped by svcntd() and widened
   f32->f64 with svcvt_f64_f32_x alone, assuming it converts the lower
   contiguous f32 lanes. Per the ARM A64 reference, SVE FCVT .s->.d reads the
   EVEN-indexed f32 lanes (source element 2*i); on any SVE register wider than
   64 bits the kernel double-counted even lanes and dropped every odd lane
   (qemu-measured relative error up to ~45% at 128-bit VL). Fixed by stepping a
   full svcntw() register and widening both halves: even via svcvt_f64_f32_x,
   odd via the SVE2 svcvtlt_f64_f32_x (FCVTLT, element 2*i+1), with merging
   adds (svadd_f64_m).

2. 2nd-moment SIMD tail squared in double (bit-exactness divergence). The
   per-row scalar tail in moment_avx2.c / moment_avx512.c / moment_neon.c
   squared in double ((double)p*(double)p) while the SIMD main loop and the
   scalar reference (moment.c: pic_ * pic_) square in float. Fixed to
   (double)(p*p). Golden-safe: the Netflix float_moment golden tests use
   576-wide content (multiple of 8/16/4), so the tail never runs.

The duplicate libvmaf.c:1881 bpc &&->|| finding is already fixed on master
(line 1903) and owned by core-engine; not touched here.

Regression coverage: new tail-only bit-exact tests in test_moment_simd.c
(test_{avx2,avx512,neon}_tail_bitexact); existing SVE2 relative-tolerance
tests cover the FCVT fix under SVE2 hardware/qemu.

Reproducer:
  meson setup build core -Denable_cuda=false -Denable_sycl=false -Denable_avx512=true
  ninja -C build && meson test -C build test_moment_simd        # 10/10 OK
  # SVE2 (no native HW): build the qemu driver in the PR notes and run
  #   qemu-aarch64-static -cpu max,sve-max-vq=1 ./sve2_check      # old: FAIL, fixed: PASS
  # golden:
  ./build/tools/vmaf --reference src01_hrc00_576x324.yuv \
    --distorted src01_hrc01_576x324.yuv --width 576 --height 324 \
    --pixel_format 420 --bitdepth 8 --feature float_moment --no_prediction \
    --output - | grep float_moment_ref2nd   # mean 4696.668388 == golden

Closes T-BUGHUNT-SIMD-2026-06-27.

+++ core/src/feature/arm64/moment_neon.c
@@ -72,8 +72,14 @@ int compute_2nd_moment_neon(const float *pic, int w, int h, int stride, double *
         }
         double tail = 0.0;
         for (; j < w; ++j) {
+            /* Square in float (not double) so the tail is bit-exact to the
+             * scalar reference (moment.c:compute_2nd_moment, `pic_ * pic_`)
+             * and to the SIMD main loop (`vmulq_f32(v, v)`).  Squaring in
+             * double here would round differently and break the ADR-0179
+             * bit-exactness contract on rows whose width is not a multiple of
+             * the SIMD lane count. */
             const float p = row[j];
-            tail += (double)p * (double)p;
+            tail += (double)(p * p);
         }
         if (tail != 0.0)
             dsum0 = vaddq_f64(dsum0, vsetq_lane_f64(tail, vdupq_n_f64(0.0), 0));
